requirements.txt
streamlit
mariadb
python-dotenv
pandas